# Conflicts:
#	gradle.properties
#	src/main/java/codes/biscuit/skyblockaddons/listeners/RenderListener.java
#	src/main/java/codes/biscuit/skyblockaddons/utils/ActionBarParser.java
#	src/main/resources/data.json
#	src/main/resources/enchants.json

+++ src/main/java/codes/biscuit/skyblockaddons/core/Attribute.java
@@ -14,9 +14,9 @@ public enum Attribute {
     MAX_FUEL(3000),
     OVERFLOW_MANA(20);
 
-    private final int defaultValue;
+    private final float defaultValue;
 
-    Attribute(int defaultValue) {
+    Attribute(float defaultValue) {
         this.defaultValue = defaultValue;
     }
 }
